requirements.txt
requests
bottle
newsapi-python
